+++ src/interpolation/polar_splines/polar_bsplines.hpp
@@ -10,7 +10,7 @@
 #include "cartesian_to_barycentric.hpp"
 #include "coord_transformation_tools.hpp"
 #include "ddc_helper.hpp"
-#include "discrete_to_cartesian.hpp"
+#include "discrete_poloidal_cs_spline_mapping.hpp"
 #include "view.hpp"
 
 namespace PolarSplines {
@@ -302,12 +302,22 @@ public:
          *                                  the new basis splines which cross the singular point.
          */
         template <class X, class Y, class SplineEvaluator, class EvalMemorySpace>
-        explicit Impl(const DiscreteToCartesian<X, Y, SplineEvaluator, R, Theta, EvalMemorySpace>&
-                              curvilinear_to_cartesian)
+        explicit Impl(const DiscretePoloidalCSSplineMapping<
+                      X,
+                      Y,
+                      SplineEvaluator,
+                      R,
+                      Theta,
+                      EvalMemorySpace>& curvilinear_to_cartesian)
         {
             static_assert(std::is_same_v<MemorySpace, Kokkos::HostSpace>);
-            using DiscreteMapping
-                    = DiscreteToCartesian<X, Y, SplineEvaluator, R, Theta, EvalMemorySpace>;
+            using DiscreteMapping = DiscretePoloidalCSSplineMapping<
+                    X,
+                    Y,
+                    SplineEvaluator,
+                    R,
+                    Theta,
+                    EvalMemorySpace>;
             static_assert(std::is_same_v<typename DiscreteMapping::BSplineR, BSplinesR>);
             static_assert(std::is_same_v<typename DiscreteMapping::BSplineTheta, BSplinesTheta>);
             using EvalExecSpace = std::conditional_t<