+++ tests/coord_transformations/refined_discrete_mapping.cpp
@@ -11,8 +11,8 @@
 #include "combined_mapping.hpp"
 #include "czarny_to_cartesian.hpp"
 #include "ddc_helper.hpp"
-#include "discrete_mapping_builder.hpp"
-#include "discrete_to_cartesian.hpp"
+#include "discrete_poloidal_cs_spline_mapping.hpp"
+#include "discrete_poloidal_cs_spline_mapping_builder.hpp"
 #include "geometry_coord_transformations_tests.hpp"
 #include "geometry_pseudo_cartesian.hpp"
 #include "inv_jacobian_o_point.hpp"
@@ -449,15 +449,19 @@ TEST(RefinedDiscreteMapping, TestRefinedDiscreteMapping)
     // Tests ---
     std::array<double, 3> results;
 
-    DiscreteToCartesianBuilder<X, Y, SplineRThetaBuilder_host, SplineRThetaEvaluator_host>
+    DiscretePoloidalCSSplineMappingBuilder<
+            X,
+            Y,
+            SplineRThetaBuilder_host,
+            SplineRThetaEvaluator_host>
             mapping_builder(
                     Kokkos::DefaultHostExecutionSpace(),
                     analytical_mapping,
                     builder,
                     spline_evaluator);
-    DiscreteToCartesian discrete_mapping = mapping_builder();
+    DiscretePoloidalCSSplineMapping discrete_mapping = mapping_builder();
 
-    RefinedDiscreteToCartesianBuilder<
+    RefinedDiscretePoloidalCSSplineMappingBuilder<
             X,
             Y,
             SplineRThetaBuilder_host,
@@ -469,9 +473,9 @@ TEST(RefinedDiscreteMapping, TestRefinedDiscreteMapping)
                     analytical_mapping,
                     builder,
                     spline_evaluator);
-    DiscreteToCartesian refined_mapping_16x32 = mapping_builder_16x32();
+    DiscretePoloidalCSSplineMapping refined_mapping_16x32 = mapping_builder_16x32();
 
-    RefinedDiscreteToCartesianBuilder<
+    RefinedDiscretePoloidalCSSplineMappingBuilder<
             X,
             Y,
             SplineRThetaBuilder_host,
@@ -483,9 +487,9 @@ TEST(RefinedDiscreteMapping, TestRefinedDiscreteMapping)
                     analytical_mapping,
                     builder,
                     spline_evaluator);
-    DiscreteToCartesian refined_mapping_32x64 = mapping_builder_32x64();
+    DiscretePoloidalCSSplineMapping refined_mapping_32x64 = mapping_builder_32x64();
 
-    RefinedDiscreteToCartesianBuilder<
+    RefinedDiscretePoloidalCSSplineMappingBuilder<
             X,
             Y,
             SplineRThetaBuilder_host,
@@ -497,7 +501,7 @@ TEST(RefinedDiscreteMapping, TestRefinedDiscreteMapping)
                     analytical_mapping,
                     builder,
                     spline_evaluator);
-    DiscreteToCartesian refined_mapping_64x128 = mapping_builder_64x128();
+    DiscretePoloidalCSSplineMapping refined_mapping_64x128 = mapping_builder_64x128();
 
 
     std::cout << std::endl